[refactor] got all tests to work again on both backend and frontend, fixed a few bugs
