+++ __tests__/jest_config.js
@@ -5,7 +5,8 @@ module.exports = {
     "^.+\\.jsx?$": "babel-jest"
   },
   bail: true,
-  verbose: false,
+  verbose: true,
+  testEnvironment: "node",
   testRegex: "/__tests__/.*.spec.(js)?$",
   globalSetup: "./jest_init.js", // This filepath is relative to the config file
 };
