+++ bin/commands/sync_models.js
@@ -5,11 +5,11 @@
 //   This is only intended to run on dev and will not work at all on production
 //
 
-const AppKernel = require('../app/AppKernel');
+const AppKernel = require('../../app/AppKernel');
 const app_kernel = new AppKernel(process.env.NODE_ENV);
 app_kernel.boot();
 
-const SyncDbSchemaCommand = require('../lib/Command/Commands/SyncDbSchemaCommand');
+const SyncDbSchemaCommand = require('../../lib/Command/Commands/SyncDbSchemaCommand');
 const command = new SyncDbSchemaCommand(app_kernel);
 
 command.run();
