+++ config/config_local.js
@@ -4,18 +4,18 @@
 module.exports = {
   port: 8000,
   db: {
-    host: '',
-    db: 'chuuni',
+    host: 'localhost',
     user: 'root',
     password: '',
-  },
-  secret: 'development_env_session_secret_lol',
-  slack: {
-    bot_user_id: '',
-    bot_token: '',
+    db: 'session_test',
   },
   nicobot: {
-    bot_user_id: '',
-    bot_token: '',
-  }
+    bot_user_id: 'U5YDC4Y21',
+    bot_token: 'xoxb-202454168069-w0qjwRsxTODGPNezbxLyyMXi',
+  },
+  slack: {},
+  fsd_workspace: {
+    legacy_token: 'xoxp-36778673491-36851507831-187811729077-e26b8ea3bd5f6cd65819ae57c092c86a',
+  },
+  secret: 'ekkusupuroooshyon',
 };
