+++ lib/FileSweeper/FileSweeper.js
@@ -27,7 +27,8 @@ class FileSweeper {
     this.handleInvalidInput = handleInvalidInput;
     this.fetchList = fetchList;
     this.token = token;
-    this.log = process.env.NODE_ENV !== 'test' ? console.log : function () {}; // to suppress logging during testing
+    this.prodEnvironment = process.env.NODE_ENV !== 'test' &&  process.env.NODE_ENV !== 'local';
+    this.log = this.prodEnvironment ? console.log : function () {}; // to suppress logging during testing
   }
 
   /**
@@ -63,7 +64,7 @@ class FileSweeper {
       this.log(item);
     });
     const self = this;
-    const prompt = process.env.NODE_ENV !== 'test' ? '\nChoose one of the options above: ' : '';
+    const prompt = this.prodEnvironment ? '\nChoose one of the options above: ' : '';
     this.rl.question(prompt, (decision) => {
       self.handleMainMenuOption(decision, eventType);
     });
@@ -128,7 +129,7 @@ class FileSweeper {
    */
   confirmDeletion(eventType, caller) {
     const self = this;
-    const prompt = process.env.NODE_ENV !== 'test' ? '\nAre you sure you want to delete the items above?\nEnter your choice (Y/N): ' : '';
+    const prompt = this.prodEnvironment ? '\nAre you sure you want to delete the items above?\nEnter your choice (Y/N): ' : '';
     this.rl.question(prompt, (option) => {
       self.handleDeletionOption(eventType, option, caller);
     });
