+++ scripts/startSweeper.js
@@ -6,10 +6,10 @@ const dependencies = [
   require('../lib/FileSweeper/utils/menuTypes'),
   new (require('../lib/FileSweeper/utils/Timer'))(),
   new (require('../lib/FileSweeper/utils/Handler'))().invalidInput,
-  new (require('../lib/FileSweeper/utils/Fetcher'))(this.httpClient).fetchList,
+  new (require('../lib/FileSweeper/utils/Fetcher'))(require('axios')).fetchList,
   require('axios'),
   (process.env.NODE_ENV === 'production' ? require('/etc/chuuni/config').fsd_workspace.legacy_token
-    : require('../config/config').fsd_workspace.legacy_token),
+    : require('../config/config_local').fsd_workspace.legacy_token),
 ];
 
 const app = new FileSweeper(...dependencies);
